feat：微信登录,查询菜品

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -58,7 +58,7 @@ public class DishController {
 
     @GetMapping("/list")
     @ApiOperation("根据分类id查询菜品")
-    public Result<List<Dish>> getDishByCategoryId(Integer categoryId) {
+    public Result<List<Dish>> getDishByCategoryId(Long categoryId) {
         log.info("根据分类id查询菜品：{}", categoryId);
         List<Dish> dishList = dishService.getDishByCategoryId(categoryId);
         return Result.success(dishList);
